% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_characteristics.R
\name{calc_starts}
\alias{calc_starts}
\title{Calculate positions of the start of \code{TRUE}-runs}
\usage{
calc_starts(x)
}
\arguments{
\item{x}{A vector. Values are coerced to logical values.}
}
\value{
The start position(s) (base-1) of \code{TRUE}-runs in \code{x}.
}
\description{
Calculate positions of the start of \code{TRUE}-runs
}
\examples{
calc_starts(c(0, 1, 1, 0, 0)) ## expected: 2
calc_starts(c(1, 1, 0, 0)) ## expected: 1
calc_starts(c(1, 1, 0, 0, 1, 1)) ## expected: 1 5
calc_starts(c(1.5, 2, 0, 0, 15, 3.5)) ## expected: 1 5
calc_starts(c(0, 0)) ## expected: numeric(0)

}
