% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{convert_precipitation}
\alias{convert_precipitation}
\title{Converts units of precipitation data}
\usage{
convert_precipitation(
  x,
  unit_from,
  unit_to = c("cm day-1", "cm month-1", "cm/day", "cm/month"),
  dpm = NA
)
}
\arguments{
\item{x}{A numeric object. Precipitation data in units of \code{unit_from}.}

\item{unit_from}{A character string. Units of data in \code{x}. Currently,
supported units include "mm/month", "mm month-1", "mm/d", "mm d-1",
"kg/m2/s", "kg m-2 s-1", "mm/s", "mm s-1", "cm/month", "cm month-1".}

\item{unit_to}{A character string. Units to which data are converted.
Currently, supported unit are "cm day-1", "cm month-1" and
respectively "cm/day", "cm/month".}

\item{dpm}{A numeric object. Number of days per month in the time series
\code{x}. Only used if one of the units is on a monthly time scale.}
}
\value{
A numeric object of the same size as \code{x} in units of
\code{unit_to}.
}
\description{
Converts units of precipitation data
}
