% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical_functions.R
\name{cor2}
\alias{cor2}
\title{Correlation function with safe-guards against errors}
\usage{
cor2(
  x,
  y = NULL,
  use = "everything",
  method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{y}{\code{NULL} (default) or a vector, matrix or data frame with
    compatible dimensions to \code{x}.   The default is equivalent to
    \code{y = x} (but more efficient).}

\item{use}{an optional character string giving a
    method for computing covariances in the presence
    of missing values.  This must be (an abbreviation of) one of the strings
    \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"},
    \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}
}
\value{
The value of \code{\link[stats]{cor}} if successful,
otherwise \code{NA}.
}
\description{
Correlation function with safe-guards against errors
}
\section{Note}{
 If \code{x} is a two-dimensional object and \code{y} is
\code{NULL} (or missing), then the correlation between the first and second
column of \code{x} is calculated -- unlike \code{\link[stats]{cor}} which
calculates all pairwise correlations among all columns of \code{x}.
}

\examples{
x <- sample(10)
x <- data.frame(
  val1 = x,
  val2 = x + stats::rnorm(10, 0, 1)
)
cor2(x)

}
