% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_characteristics.R
\name{count_peaks}
\alias{count_peaks}
\title{Count the number of peaks in a series of values}
\usage{
count_peaks(x, min_change = 0)
}
\arguments{
\item{x}{A numeric vector}

\item{min_change}{A numeric value. The size of a minimal change between
two consecutive values of \code{x} to count as increase/decrease.}
}
\value{
An integer value representing the number of peaks
}
\description{
Peaks are identified as patterns of increase, possibly followed by a
constant stretch, which ends in a decrease.
}
\examples{
count_peaks(c(0, 1, 0, 1, 0)) ## expect 2 peaks
count_peaks(c(0, 1, 0, 1)) ## expect 1 peak
count_peaks(c(1, 0, 1)) ## expect 0 peak
count_peaks(c(0, 1, 1, 2.5, 5.1, 4.9)) ## expect 1 peak
count_peaks(c(0, 1, 1, 2.5, 5.1, 4.9), min_change = 0.5) ## expect 0 peak
count_peaks(c(0, 1, 1, 0.8, 5.1, 4)) ## expect 2 peak
count_peaks(c(0, 1, 1, 0.8, 5.1, 4), min_change = 0.1) ## expect 2 peak
count_peaks(c(0, 1, 1, 0.8, 5.1, 4), min_change = 0.5) ## expect 1 peak

}
