% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous_functions.R
\name{extend_range}
\alias{extend_range}
\title{Extend range by factor}
\usage{
extend_range(x, fact = 1, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{fact}{A numeric value}

\item{na.rm}{A logical value}
}
\description{
Extend range by factor
}
