% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit.R
\name{gof}
\alias{gof}
\alias{rmse}
\alias{mae}
\alias{NSE}
\title{Goodness of fit measures between observations and simulated values}
\usage{
rmse(obs, sim, na.rm = FALSE)

mae(obs, sim, na.rm = FALSE)

NSE(obs, sim, na.rm = FALSE)
}
\arguments{
\item{obs}{A numeric one- or two-dimensional object.}

\item{sim}{A numeric one- or two-dimensional object.}

\item{na.rm}{A logical value. Should (pairwise) missing values be removed?}
}
\value{
A numeric vector of length corresponding to the number of columns
  (which is one in case of one-dimensional inputs.)
}
\description{
Goodness of fit measures between observations and simulated values
}
\section{Functions}{
\itemize{
\item \code{rmse}: Root Mean Square Error

\item \code{mae}: Mean Absolute Error

\item \code{NSE}: Nash-Sutcliffe efficiency
}}

\section{Details}{
 \code{obs} and \code{sim} must have the same dimensions.
  If they are two-dimensional objects, then the GOFs are calculated for
  each column separately.
}

\references{
Nash, J. E. and Sutcliffe, J. V. (1970). River flow forecasting
  through conceptual models part I: A discussion of principles.
  Journal of Hydrology, 10:3, 282-290.
}
