% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounded_values.R
\name{handle_NAs}
\alias{handle_NAs}
\title{Apply \code{napredict} carefully}
\usage{
handle_NAs(x, na.index, na.act)
}
\arguments{
\item{x}{a vector, data frame, or matrix to be adjusted based upon the
    missing value information.}

\item{na.index}{An object that has a length greater than 0 if \code{NA}s
are present in \code{x}.}

\item{na.act}{An object produced by a
\code{\link[stats]{na.action}} function.}
}
\description{
Apply \code{napredict} carefully
}
\examples{
x <- c(3, NA, 2, NA, 4)
# Compare
cumsum(x)

x2 <- stats::na.exclude(x)
handle_NAs(cumsum(x2), anyNA(x), attr(x2, "na.action"))

}
\seealso{
\code{\link[stats]{napredict}}
}
