% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootfinding_optimization.R
\name{has_uniroots}
\alias{has_uniroots}
\title{Determine whether result(s) of \code{uniroots} was successful}
\usage{
has_uniroots(x, tol = .Machine$double.eps^0.25)
}
\arguments{
\item{x}{An object returned from \code{uniroots}}

\item{tol}{A numeric value. The convergence tolerance.}
}
\value{
A logical value or a vector of logical values.
}
\description{
Determine whether result(s) of \code{uniroots} was successful
}
\examples{
r1 <- uniroots(
  f = function(x) -1 + x ^ 2,
  xlim = c(-10, 10),
  expected_nroots = 2
)

has_uniroots(r1)

r2 <- uniroots(f = function(x) -1 + x ^ 2, xlim = c(-10, -5))
has_uniroots(r2)

r3 <- uniroots(f = function(x) Inf, xlim = c(-10, -5))
has_uniroots(r3)

}
