% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{isLeapYear}
\alias{isLeapYear}
\title{Check whether a Gregorian calendar year is a leap year}
\usage{
isLeapYear(y)
}
\arguments{
\item{y}{An integer value. A Gregorian calendar year.}
}
\description{
Check whether a Gregorian calendar year is a leap year
}
\examples{
isFALSE(isLeapYear(1900))
isTRUE(isLeapYear(1980))
isTRUE(isLeapYear(2000))

}
