% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous_functions.R
\name{make_chunks}
\alias{make_chunks}
\title{Divide number of tasks into chunks}
\usage{
make_chunks(nx, n_chunks, chunk_size = 1000L)
}
\arguments{
\item{nx}{An integer value. Number of tasks.}

\item{n_chunks}{An integer value or missing. The number of chunks.}

\item{chunk_size}{An integer value or missing. The number of tasks per chunk.}
}
\value{
A list of integer vectors (chunks) that contain indices of
  all \code{nx} tasks; either, the number of chunks is \code{<= n_chunks},
  or the number of tasks/indices per chunk is \code{<= chunk_size}.
}
\description{
Create a list that contains chunks of indices
(that reference tasks by position) of either \itemize{
  \item a specified number of chunks \code{n_chunks},
    useful, e.g., if we have a fixed number of processing units, or
  \item a specified number of tasks per chunk \code{chunk_size},
    useful, e.g., if we have a (upper) fixed number of elements than can be
    processed at a time.
}
The
}
\examples{
nx <- 16

## Create list of indices by number of chunks,
(ids <- make_chunks(nx, n_chunks = 3))
print(length(ids))

## Create list of indices by size per chunk,
(ids <- make_chunks(nx, chunk_size = 5))
print(lengths(ids))

}
\seealso{
\code{\link[parallel]{splitIndices}}
}
