% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_characteristics.R
\name{max_duration}
\alias{max_duration}
\title{Determine the length of the (first) longest run}
\usage{
max_duration(x, target_val = 1L, return_doys = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{target_val}{A value. Runs of \code{target_val} in \code{x} are used.}

\item{return_doys}{A logical value.}
}
\value{
A numeric vector with one or three elements including
the length of the longest run of value \code{target_val}
(the first if multiple runs are of equally the longest) and,
if \code{return_doys == TRUE}, then the start and end
position of that longest run.
}
\description{
Determine the length of the (first) longest run
}
\examples{
max_duration(0, target_val = 1)
x <- c(rep(3, 3), rep(10, 10), 4, rep(10, 20), rep(3, 6))
max_duration(x, 10, FALSE)
max_duration(x, 10, TRUE)

}
