% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounded_values.R
\name{oob}
\alias{oob}
\alias{squash_into_low_high}
\alias{cut0Inf}
\alias{finite01}
\title{Replace values below or above a threshold (or bound) with new values}
\usage{
squash_into_low_high(
  x,
  val_low = 0,
  val_low_replace = val_low,
  val_high = 1,
  val_high_replace = val_high
)

cut0Inf(x, val = NA)

finite01(x, val_low_replace = 0, val_high_replace = 1)
}
\arguments{
\item{x}{A vector.}

\item{val_low}{A value. The lower threshold (or bound).}

\item{val_low_replace}{A value. The replacement for
values in \code{x} below \code{val_low}.}

\item{val_high}{A value. The upper threshold (or bound).}

\item{val_high_replace}{A value. The replacement for
values in \code{x} below \code{val_low}.}

\item{val}{A value. The replacement value.}
}
\description{
Replace values below or above a threshold (or bound) with new values
}
\examples{
x <- seq(-5, 5)
squash_into_low_high(x)
squash_into_low_high(x, val_low = 0, val_high = 3)
squash_into_low_high(
  x,
  val_low = 0, val_low_replace = -999,
  val_high = 3, val_high_replace = 999
)

cut0Inf(c(-5, 0, NA, 0.5, 1, 2, Inf))

finite01(c(-5, 0, NA, 0.5, 1, 2, Inf))

}
