% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSW2utils-package.R
\docType{package}
\name{rSW2utils}
\alias{rSW2utils}
\alias{rSW2utils-package}
\title{Package \pkg{rSW2wutils}: summary information}
\description{
Collection of general purpose functions to support SOILWAT2/STEPWAT2 simulation experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DrylandEcology/rSW2utils}
  \item Report bugs at \url{https://github.com/DrylandEcology/rSW2utils/issues}
}

}
\author{
\strong{Maintainer}: Daniel Schlaepfer \email{danielrschlaepfer@gmail.com} (\href{https://orcid.org/0000-0001-9973-2065}{ORCID})

}
