% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_characteristics.R
\name{scale_by_sum}
\alias{scale_by_sum}
\title{Scale values of a vector so that the result sums to 1}
\usage{
scale_by_sum(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Scale values of a vector so that the result sums to 1
}
\examples{
x <- c(1:6, 8)

# Equivalent but more efficient than `scale_to_reference_fun`
scale_by_sum(x)
scale_to_reference_fun(x, 1, fun = sum)

}
\seealso{
\code{\link{scale_to_reference_fun}}
}
