% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_characteristics.R
\name{scale_rounded_by_sum}
\alias{scale_rounded_by_sum}
\title{Scale values so that the rounded values sum to 1}
\usage{
scale_rounded_by_sum(
  x,
  digits,
  icolumn_adjust = 1L,
  max_iter = max(4L, digits + 1L),
  verbose = FALSE
)
}
\arguments{
\item{x}{A one or two-dimensional numeric object with positive values:
a vector, data frame, matrix, or array.
A vector is treated as a 1-row matrix.}

\item{digits}{A integer value. The number of decimal places to round;
a number between one and four.}

\item{icolumn_adjust}{An integer value. The column number of
\code{x} that will be adjusted if rounded values have not converged
to sum to 1 after \code{max_iter}.}

\item{max_iter}{An integer value. The maximum number of iterations.}

\item{verbose}{A logical value. Reports a warning if \code{TRUE} and
\code{max_iter} is exceeded.}
}
\description{
Scale values so that the rounded values sum to 1
}
\section{Notes}{

This function may fail due to integer division by zero.


This function may also not converge to round values so that they sum to 1,
particularly if too few \code{digits} are available to round. In that case,
attempt to increase \code{digits}. However, the return values should sum
to within \code{digits} of 1.
}

\examples{
x <- c(1:6, 8) / 10
scale_rounded_by_sum(x, digits = 1, icolumn_adjust = 7)
scale_rounded_by_sum(x, digits = 2, icolumn_adjust = 7)
scale_rounded_by_sum(x, digits = 4, icolumn_adjust = 7)

x <- matrix(1:12, nrow = 4, ncol = 3)
scale_rounded_by_sum(x, digits = 1)
scale_rounded_by_sum(x, digits = 2)
scale_rounded_by_sum(x, digits = 4)

}
\seealso{
\code{\link{scale_by_sum}}, \code{\link[base]{round}}
}
