% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_characteristics.R
\name{scale_to_reference_fun}
\alias{scale_to_reference_fun}
\title{Scale values so that they retain the value of a function
applied to reference values}
\usage{
scale_to_reference_fun(x, x_ref, fun, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector. Values to scale.}

\item{x_ref}{Numeric vector.
Values from which the reference value is calculated.}

\item{fun}{A function or character vector identifying a function.}

\item{na.rm}{A logical value. Argument is passed to \code{fun}.}
}
\value{
A numeric vector of the same length as \code{x}.
}
\description{
Scale values so that they retain the value of a function
applied to reference values
}
\examples{
x <- c(1:6, 8)
x0 <- c(10, -1)
x0NA <- c(x0, NA)

# Scale to retain value of reference maximum
scale_to_reference_fun(x, x0, fun = max)

# Scale to retain value of reference sum
scale_to_reference_fun(x, x0, fun = sum)

# Scale to retain value of reference mean
scale_to_reference_fun(x, x0, fun = mean)

# Scale to retain value of reference length (mostly nonsensical!)
scale_to_reference_fun(x, x0, fun = length)

# Scale and handle NAs
scale_to_reference_fun(x, x0NA, fun = mean, na.rm = FALSE)
scale_to_reference_fun(x, x0NA, fun = max, na.rm = TRUE)

}
\seealso{
\code{\link{scale_to_reference_peak_frequency}}
}
