% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical_functions.R
\name{scale_by_sum}
\alias{scale_by_sum}
\title{Scale values of a vector so that the result sums to 1}
\usage{
scale_by_sum(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Scale values of a vector so that the result sums to 1
}
